# design/03-platform/openapi.yaml
openapi: 3.1.0
info:
  title: Connectome API
  version: 1.0.0-design
  description: Core of the public API. All model writes go through changes. Edit shapes mirror 05-structures/changes.ts.
servers:
  - url: https://{region}.connectome.app/api/v1
    variables:
      region: { default: eu, enum: [eu, us, au] }
security:
  - bearer: []
paths:
...
  /repositories/{repo}/changes:
    parameters: [ { $ref: '#/components/parameters/repo' }, { $ref: '#/components/parameters/scenario' } ]
    post:
      summary: Commit a change (atomic list of edits)
      parameters:
        - { name: preview, in: query, schema: { type: boolean, default: false } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChangeRequestBody' } } }
      responses:
        '200': { description: 'Preview (preview=true): the outcome if committed; nothing is written. seq is null', content: { application/json: { schema: { $ref: '#/components/schemas/ChangeResult' } } } }
        '201': { description: 'Committed. Resending a change id returns the original outcome', content: { application/json: { schema: { $ref: '#/components/schemas/ChangeResult' } } } }
        '202': { description: Held for review by a governance policy, content: { application/json: { schema: { $ref: '#/components/schemas/ChangeResult' } } } }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    get:
      summary: Committed changes since a sequence number
      parameters: [ { name: since, in: query, required: true, schema: { type: integer } } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ChangeSummary' } } } } }
  /repositories/{repo}/changes/{changeId}/undo:
    parameters: [ { $ref: '#/components/parameters/repo' }, { name: changeId, in: path, required: true, schema: { type: string } } ]
    post:
      summary: Undo one of your own changes (a new change built from inverse edits). 403 for other people's changes, 422 if already undone
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, properties: { id: { type: string, description: 'Id for the new change, so retries are safe' } } } } }
      responses:
        '201': { description: Undone, content: { application/json: { schema: { $ref: '#/components/schemas/ChangeResult' } } } }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
  parameters:
    repo: { name: repo, in: path, required: true, schema: { type: string } }
...
    scenario: { name: scenario, in: query, required: false, schema: { type: string }, description: Scenario id (default baseline) }
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              title: { type: string }
              status: { type: integer }
              code: { type: string, enum: [conflict, ruleViolation, forbidden, notFound, invalid, gone, unauthenticated] }
              details:
                type: array
                items: { type: object, properties: { editIndex: { type: integer }, property: { type: string }, rule: { type: string }, message: { type: string }, itemId: { type: string }, changedBy: { type: string } } }
  schemas:
...
    Edit:
      type: object
      required: [edit]
      description: One edit; see 05-structures/changes.ts for every shape.
      properties:
        edit:
          type: string
          enum: [createObject, setProperties, renameObject, moveToFolder, changeObjectType, deleteObject,
                 createRelationship, reconnectRelationship, deleteRelationship,
                 createFolder, renameFolder, moveFolder, deleteFolder,
                 createDiagram, updateDiagram, deleteDiagram,
                 addObjectOccurrence, moveObjectOccurrence, styleOccurrence, removeOccurrence,
                 addRelationshipOccurrence, routeRelationshipOccurrence, addAnnotation, updateAnnotation, removeAnnotation,
                 setTags]
      additionalProperties: true
    ChangeRequestBody:
      type: object
      required: [id, label, edits]
      properties:
        id: { type: string }
        label: { type: string }
        changeRequestId: { type: string }
        edits: { type: array, minItems: 1, maxItems: 10000, items: { $ref: '#/components/schemas/Edit' } }
    ChangeResult:
      type: object
      properties:
        seq: { type: [integer, 'null'] }
        changeId: { type: string }
        preview: { type: boolean }
        versions: { type: object, additionalProperties: { type: integer } }
        heldInChangeRequest: { type: [string, 'null'] }
        findings:
          type: array
          items: { type: object, properties: { rule: { type: string }, itemId: { type: string }, severity: { type: string }, message: { type: string } } }
    ChangeSummary:
      type: object
      properties:
        seq: { type: integer }
        changeId: { type: string }
        scenarioId: { type: string }
        label: { type: string }
        actor: { type: [string, 'null'] }
        source: { type: string }
        committedAt: { type: string, format: date-time }
        edits: { type: array, items: { $ref: '#/components/schemas/Edit' } }
...